+++ rs/tests/financial_integrations/icrc1_agent_test.rs
@@ -1,15 +1,14 @@
-mod icrc1_agent_test;
+mod icrc1_agent_tests;
 
 use anyhow::Result;
 
 use ic_system_test_driver::driver::group::SystemTestGroup;
 use ic_system_test_driver::systest;
-use icrc1_agent_test;
 
 fn main() -> Result<()> {
     SystemTestGroup::new()
-        .with_setup(icrc1_agent_test::config)
-        .add_test(systest!(icrc1_agent_test::test))
+        .with_setup(icrc1_agent_tests::config)
+        .add_test(systest!(icrc1_agent_tests::test))
         .execute_from_args()?;
     Ok(())
 }
